+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## To Be Released
+## 8.2.0
 
 * feat(projects): add edit project event
 * feat(projects): add delete project event